requirements.txt
requests
bs4
lxml
sqlalchemy
psycopg2-binary
python-decouple
